PID_2.py
this is the mostly completed code from the barebones. I'm not sure how to code the integral component of the code. I left the sensor values as array components as they will be constantly imported from the sensors and calculated when in the embedded system. feel free to edit this code and make more progress on it.